+++ src/main/kotlin/com/abishake/fitnesstracker/service/ExerciseService.kt
@@ -2,7 +2,6 @@ package com.abishake.fitnesstracker.service
 
 import com.abishake.fitnesstracker.models.Exercise
 import com.abishake.fitnesstracker.repositories.ExerciseRepository
-import org.springframework.dao.DataIntegrityViolationException
 import org.springframework.stereotype.Service
 
 @Service
@@ -10,17 +9,16 @@ class ExerciseService(
     private val exerciseRepository: ExerciseRepository
 ) {
 
-    private val className = ExerciseService::class
-
     // CREATE
-    fun createExercise(name: String, description: String): Exercise {
+    fun createExercise(name: String, description: String): List<Exercise> {
         val existingEntry = exerciseRepository.findByNameAndDescription(name, description)
         return if (existingEntry.isEmpty()) {
-            exerciseRepository.saveAndFlush(
+            val saved = exerciseRepository.saveAndFlush(
                 Exercise(null, name, description)
             )
+            return listOf(saved)
         } else {
-            existingEntry[0]
+            listOf(existingEntry.first())
         }
     }
 
@@ -29,11 +27,11 @@ class ExerciseService(
         return exerciseRepository.findAll()
     }
 
-    fun getExerciseById(id: Long): Exercise {
-        if (exerciseRepository.findById(id).isPresent) {
-            return exerciseRepository.findById(id).get()
+    fun getExerciseById(id: Long): List<Exercise> {
+        return if (exerciseRepository.findById(id).isPresent) {
+            listOf(exerciseRepository.findById(id).get())
         } else {
-            throw Exception("Exercise Id = $id not found for getExerciseById() >> $className")
+            listOf()
         }
     }
 
@@ -42,23 +40,22 @@ class ExerciseService(
     }
 
     //  UPDATE
-    fun updateExerciseById(id: Long, name: String, description: String): Exercise {
-        try {
-            val exercise = getExerciseById(id)
-            exercise.name = name
-            exercise.description = description
+    fun updateExerciseById(id: Long, name: String, description: String): List<Exercise> {
+        val exerciseInList = getExerciseById(id)
+        val exercise = exerciseInList.first()
+        exercise.name = name
+        exercise.description = description
 
-            exerciseRepository.saveAndFlush(exercise)
-            return exercise
-        } catch (e: Exception) {
-            throw Exception("Exception in updateExerciseById() >> $className \n $e", e)
-        }
+        val saved = exerciseRepository.saveAndFlush(exercise)
+        return listOf(saved)
     }
 
     // DELETE
-    fun deleteExerciseById(id: Long): Exercise {
+    fun deleteExerciseById(id: Long): List<Exercise> {
         val exercise = getExerciseById(id)
-        exerciseRepository.deleteById(id)
+        if (exercise.isNotEmpty()) {
+            exerciseRepository.deleteById(id)
+        }
         return exercise
     }
 }
